requirements.txt
fastapi==0.109.2
uvicorn==0.25.0
supabase==2.3.0
openai==1.13.3
python-dotenv==1.0.0
httpx==0.25.2
jinja2==3.1.2
pydantic==2.5.2
pytest==7.4.3
pytest-asyncio==0.21.1
newsapi-python==0.2.7
python-multipart==0.0.6
email-validator==2.1.0 